Try fix

+++ 3.1.0/src/google/protobuf/stubs/atomicops.h
@@ -86,8 +86,8 @@ typedef int32 Atomic32;
 typedef intptr_t Atomic64;
 #endif
 #else
-typedef int32 Atomic32;
-typedef intptr_t Atomic64;
+typedef intptr_t Atomic32;
+typedef int64 Atomic64;
 #endif
 #endif
